requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto
absl-py==0.7.1
astor==0.8.0
boto==2.49.0
boto3==1.9.156
botocore==1.12.156
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
gast==0.2.2
gensim==3.7.3
grpcio==1.21.1
h5py==2.9.0
idna==2.8
jieba==0.39
jmespath==0.9.4
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
mock==3.0.5
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
s3transfer==0.2.0
# scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
smart-open==1.8.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
xlrd==1.2.0
gunicorn